HS110-interface/requirements.txt
prometheus-client==0.9.0
python_kasa==0.4.0.dev0
redis==3.5.3
apscheduler==3.6.3